# Fix type instability in sort for few columns case and fix issorted bug