--- a/src/abstractdataframe/sort.jl
+++ b/src/abstractdataframe/sort.jl
@@ -120,15 +120,44 @@ col_ordering(o::DFPerm{O}, i::Int) where {O<:Ordering} = o.ord
 col_ordering(o::DFPerm{V}, i::Int) where {V<:AbstractVector} = o.ord[i]
 
 function Sort.lt(o::DFPerm, a, b)
-    @inbounds for i in 1:length(o.cols)
-        ord = col_ordering(o, i)
-        col = o.cols[i]
+    @inbounds begin
+        loc = length(o.cols)
+        loc < 1 && return false
+
+        ord = col_ordering(o, 1)
+        col = o.cols[1]
         va = col[a]
         vb = col[b]
         lt(ord, va, vb) && return true
         lt(ord, vb, va) && return false
+        loc == 1 && return false
+
+        ord = col_ordering(o, 1)
+        col = o.cols[2]
+        va = col[a]
+        vb = col[b]
+        lt(ord, va, vb) && return true
+        lt(ord, vb, va) && return false
+        loc == 2 && return false
+
+        ord = col_ordering(o, 1)
+        col = o.cols[3]
+        va = col[a]
+        vb = col[b]
+        lt(ord, va, vb) && return true
+        lt(ord, vb, va) && return false
+        loc == 3 && return false
+
+        for i in 4:length(o.cols)
+            ord = col_ordering(o, i)
+            col = o.cols[i]
+            va = col[a]
+            vb = col[b]
+            lt(ord, va, vb) && return true
+            lt(ord, vb, va) && return false
+        end
     end
-    false # a and b are equal
+    return false # a and b are equal
 end
 
 ###