--- a/test/sort.jl
+++ b/test/sort.jl
@@ -160,4 +160,29 @@ end
     @test sort(view(df, 1:2, 1:2), view=true) == sort(view(df, 1:2, 1:2))
 end
 
+@testset "hard tests of different sorting orders" begin
+    Random.seed!(1234)
+    df = DataFrame(rand([0, 1], 10^5, 3), :auto)
+    @test sortperm(df, :x1) == sortperm(df.x1)
+    @test sortperm(df, [:x1, :x2]) == sortperm(tuple.(df.x1, df.x2))
+    @test sortperm(df, [:x1, :x2, :x3]) == sortperm(tuple.(df.x1, df.x2, df.x3))
+    @test sortperm(df, :x1, rev=true) == sortperm(df.x1, rev=true)
+    @test sortperm(df, [:x1, :x2], rev=true) ==
+          sortperm(tuple.(df.x1, df.x2), rev=true)
+    @test sortperm(df, [:x1, :x2, :x3], rev=true) ==
+          sortperm(tuple.(df.x1, df.x2, df.x3), rev=true)
+
+    for r1 in (true, false)
+        @test sortperm(df, order(:x1, rev=r1)) == sortperm((1 - 2*r1) * df.x1)
+        for r2 in (true, false)
+            @test sortperm(df, [order(:x1, rev=r1), order(:x2, rev=r2)]) ==
+                  sortperm(tuple.((1 - 2*r1) * df.x1, (1 - 2*r2) * df.x2))
+            for r3 in (true, false)
+                @test sortperm(df, [order(:x1, rev=r1), order(:x2, rev=r2), order(:x3, rev=r3)]) ==
+                      sortperm(tuple.((1 - 2*r1) * df.x1, (1 - 2*r2) * df.x2, (1 - 2*r3) * df.x3))
+            end
+        end
+    end
+end
+
 end # module