Fixes https://github.com/JuliaData/DataFrames.jl/issues/2745 by loop unrolling.

Probably is is possible to do it in a smarter way using metaprogramming (but maybe not) so I am marking it as a draft.