@bkamins 2021-05-02
We have a bad design of sorting in general:
```
julia> df = DataFrame(x=rand(10^7), y=rand(10^7));

julia> @time sort(df, [:x, :y]);
  5.751711 seconds (1.03 M allocations: 1021.159 MiB, 4.52% gc time)

julia> @time sort(df, [:x, order(:y, rev=true)]);
 18.962380 seconds (695.47 M allocations: 11.345 GiB, 5.67% gc time)
```

I will push an update (and then the code will be simplified with recursion)