@bkamins 2021-05-03
Timings after this PR:
```
julia> df = DataFrame(x=rand(10^7), y=rand(10^7));

julia> @btime sort($df, [:x, :y]);
  5.501 s (1029745 allocations: 1021.14 MiB)

julia> @btime sort($df, [:x, order(:y, rev=true)]);
  5.426 s (1029758 allocations: 1021.14 MiB)
```
and
```
julia> function mwedates()
         #build the sample
         dts = reduce(vcat, [[Date(2011,11,11) + Day(i) for j in 1:10^4] for i in 1:100])
         mdts = dts |> Vector{Union{Date, Missing}}
         id = reduce(vcat, [[j for j in 1:10^4] for i in 1:100])
         df = DataFrame(date=dts, mdate = mdts, id=id)

         #shuffle
         df = df[randperm(10^6), :]

         print("sort date and id: ")
         @btime sort($df, [:date, :id])
         print("sort date(with missings) and id: ")
         @btime sort($df, [:mdate, :id])
         print("work around performance: ")
         @btime begin
           $df.mdateconverted = $df.mdate |> Vector{Date}
           sort($df, [:mdateconverted, :id])
         end
       end
mwedates (generic function with 1 method)

julia> mwedates()
sort date and id:   260.934 ms (64997 allocations: 92.79 MiB)
sort date(with missings) and id:   329.115 ms (64997 allocations: 92.79 MiB)
work around performance:   272.197 ms (65004 allocations: 108.05 MiB)
```
so all is OK (the penalty of `missing` has to be accepted I think as work-around uses knowledge of the data)