@clintonTE 2021-05-03
> so all is OK (the penalty of `missing` has to be accepted I think as work-around uses knowledge of the data)

Yeah, that little overhead is fantastic. Thank you!