@bkamins 2021-05-03
> Not sure whether that would be fast.

This would be slower AFAICT, because now we are short circuting (so mostly only one comparison is needed on the first column in typical cases)