@bkamins 2021-05-03
We cannot use recursion, as it will break on very wide data frames. I will fix it tomorrow.