@bkamins 2021-05-04
> When sorting on bitstype columns which can use optimized algorithms

I agree, but currently we have:
```
julia> using DataFrames

julia> using BenchmarkTools

julia> df = DataFrame(rand(1:10, 10^6, 2), :auto);

julia> @btime sort($df, :x1);
  82.997 ms (239220 allocations: 182.06 MiB)

julia> @btime sort($df, [:x1, :x2]);
  126.545 ms (239142 allocations: 170.64 MiB)

julia> @btime sort($df.x1);
  2.694 ms (3 allocations: 7.63 MiB)

julia> t = collect(zip(df.x1, df.x2)); @btime sort($t);
  65.424 ms (4 allocations: 22.89 MiB)
```
so I think that the first step should be to make sorting on single column fast (which we do at some point - in general sorting and reshaping are things to work on in the near future as these were the areas here no new things were added for a long time). But for this PR I would concentrate the design on fixing type instability issues.