@bkamins 2021-05-04
Here are the benchmarks after the fix of recursion.
The conclusion is - in general we are faster, unless someone is sorting a very wide table on all columns (or in general sorting on very many columns - in this case we are slower although I use essentially the same code as previously - so it seems the compiler is not able to optimize things correctly in this case yet; I could fix it, but I think it is not worth it as sorting on super many columns is not very useful anyway).

Additionally I have discovered a bug in `issorted` that is fixed now.

this PR
```
julia> using DataFrames, Random, StatsBase, Dates, BenchmarkTools

julia> Random.seed!(1234)
MersenneTwister(1234)

julia> df = DataFrame(x=rand(10^6), y=rand(10^6));

julia> @time sort(df, [:x, :y]);
  1.554144 seconds (3.50 M allocations: 274.667 MiB, 3.56% gc time, 80.51% compilation time)

julia> @btime sort($df, [:x, :y]);
  275.927 ms (64998 allocations: 84.21 MiB)

julia> @time sort(df, [:x, order(:y, rev=true)]);
  0.823464 seconds (769.23 k allocations: 123.305 MiB, 1.69% gc time, 65.93% compilation time)

julia> @btime sort($df, [:x, order(:y, rev=true)]);
  279.058 ms (65005 allocations: 84.21 MiB)

julia> function mwedates()
                #build the sample
                dts = reduce(vcat, [[Date(2011,11,11) + Day(i) for j in 1:10^4] for i in 1:100])
                mdts = dts |> Vector{Union{Date, Missing}}
                id = reduce(vcat, [[j for j in 1:10^4] for i in 1:100])
                df = DataFrame(date=dts, mdate = mdts, id=id)

                #shuffle
                df = df[randperm(10^6), :]

                print("sort date and id: ")
                @btime sort($df, [:date, :id])
                print("sort date(with missings) and id: ")
                @btime sort($df, [:mdate, :id])
                print("work around performance: ")
                @btime begin
                  $df.mdateconverted = $df.mdate |> Vector{Date}
                  sort($df, [:mdateconverted, :id])
                end
              end
mwedates (generic function with 1 method)

julia> mwedates();
sort date and id:   276.888 ms (64997 allocations: 92.79 MiB)
sort date(with missings) and id:   345.657 ms (64997 allocations: 92.79 MiB)
work around performance:   279.926 ms (65005 allocations: 108.05 MiB)

julia> df = DataFrame(ones(10,1000), :auto);

julia> @time sort(df);
  0.603286 seconds (156.87 k allocations: 9.166 MiB, 99.74% compilation time)

julia> @btime sort(df);
  364.711 μs (6400 allocations: 322.86 KiB)

julia> df = DataFrame(ones(Int, 10000, 100), :auto);

julia> @time sort(df);
  0.502271 seconds (403.55 k allocations: 29.421 MiB, 1.04% gc time, 95.94% compilation time)

julia> @btime sort(df);
  15.511 ms (326 allocations: 7.72 MiB)

julia> df = DataFrame(ones(Bool, 100000, 15), :auto);

julia> @time sort(df);
  0.607865 seconds (735.55 k allocations: 43.240 MiB, 99.39% compilation time)

julia> @btime sort(df);
  3.011 ms (71 allocations: 2.20 MiB)
```

current `main`
```
julia> using DataFrames, Random, StatsBase, Dates, BenchmarkTools

julia> Random.seed!(1234)
MersenneTwister(1234)

julia> df = DataFrame(x=rand(10^6), y=rand(10^6));

julia> @time sort(df, [:x, :y]);
  1.624158 seconds (3.52 M allocations: 276.283 MiB, 7.45% gc time, 81.75% compilation time)

julia> @btime sort($df, [:x, :y]);
  272.999 ms (64998 allocations: 84.21 MiB)

julia> @time sort(df, [:x, order(:y, rev=true)]);
  1.499732 seconds (58.89 M allocations: 1008.693 MiB, 3.29% gc time, 33.07% compilation time)

julia> @btime sort($df, [:x, order(:y, rev=true)]);
  1.031 s (58215802 allocations: 971.52 MiB)

julia> function mwedates()
                #build the sample
                dts = reduce(vcat, [[Date(2011,11,11) + Day(i) for j in 1:10^4] for i in 1:100])
                mdts = dts |> Vector{Union{Date, Missing}}
                id = reduce(vcat, [[j for j in 1:10^4] for i in 1:100])
                df = DataFrame(date=dts, mdate = mdts, id=id)

                #shuffle
                df = df[randperm(10^6), :]

                print("sort date and id: ")
                @btime sort($df, [:date, :id])
                print("sort date(with missings) and id: ")
                @btime sort($df, [:mdate, :id])
                print("work around performance: ")
                @btime begin
                  $df.mdateconverted = $df.mdate |> Vector{Date}
                  sort($df, [:mdateconverted, :id])
                end
              end
mwedates (generic function with 1 method)

julia> mwedates();
sort date and id:   379.087 ms (64997 allocations: 92.79 MiB)
sort date(with missings) and id:   2.419 s (105424568 allocations: 1.66 GiB)
work around performance:   385.658 ms (65005 allocations: 108.05 MiB)

julia> df = DataFrame(ones(10,1000), :auto);

julia> @time sort(df);
  0.606793 seconds (152.38 k allocations: 9.070 MiB, 99.87% compilation time)

julia> @btime sort(df);
  258.662 μs (1999 allocations: 254.09 KiB)

julia> df = DataFrame(ones(Int, 10000, 100), :auto);

julia> @time sort(df);
  0.431671 seconds (419.29 k allocations: 30.190 MiB, 1.32% gc time, 98.73% compilation time)

julia> @btime sort(df);
  3.117 ms (326 allocations: 7.72 MiB)

julia> df = DataFrame(ones(Bool, 100000, 15), :auto);

julia> @time sort(df);
  0.452559 seconds (474.43 k allocations: 27.729 MiB, 1.04% gc time, 99.14% compilation time)

julia> @btime sort(df);
  3.281 ms (74 allocations: 2.20 MiB)
```